requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
